requirements.txt
pandas
textstat
sentence-transformers
scikit-learn
transformers
torch
plotly
openai
python-dotenv
